+++ launch/navsat_transform.launch.py
@@ -21,14 +21,23 @@ from launch.substitutions import EnvironmentVariable
 import pathlib
 import launch.actions
 from launch.actions import DeclareLaunchArgument
+from launch.substitutions import LaunchConfiguration
 
 def generate_launch_description():
+    autostart = DeclareLaunchArgument(
+        'autostart',
+        default_value='false',
+        description='Automatically configure and activate the node. Set to false for managed lifecycle control.')
+
     return LaunchDescription([
-        launch_ros.actions.Node(
+        autostart,
+        launch_ros.actions.LifecycleNode(
             package='robot_localization',
             executable='navsat_transform_node',
             name='navsat_transform_node',
+            namespace='',
             output='screen',
+            autostart=LaunchConfiguration('autostart'),
             parameters=[os.path.join(get_package_share_directory("robot_localization"), 'params', 'navsat_transform.yaml')],
            ),
 ])